requirements.txt
langchain[groq]
python-dotenv
google-generativeai
langgraph
langchain-google-genai
python-docx
matplotlib
networkx